==> ddm_analysis/step3_Inference_testing_study_1.ipynb <==
import ipyparallel
import hddm
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline
import seaborn as sns
print(hddm.__version__)

# Confirmatory inference testing

# load the HDDM model fitted in step XX
m = hddm.load("m_full_hddm_study1_new_subj10_removed")

# get the MCMC chains of HDDM model
v_0V0A, v_0V1A, v_1V0A, v_1V1A, v_1Vne1A = m.nodes_db.loc[['v(0V0A)', 'v(0V1A)', 'v(1V0A)', 'v(1V1A)', 'v(1Vne1A)'], 'node']

# save the parameter samples for plotting
pd.DataFrame({
    "v_0V0A": v_0V0A.trace(),
    "v_0V1A":v_0V1A.trace(),
    "v_1V0A":v_1V0A.trace(),
    "v_1V1A":v_1V1A.trace(),
    "v_1Vne1A":v_1Vne1A.trace(),
             }).to_csv("plot/study_1_ddm_confirmative_analysis.csv")

# test the percentage of posteriors higher and lower than 0
def inference_test(trace):
    probability1 = round((trace < 0).mean(), 3) 
    probability2 = round((1-probability1), 3) 
    return "{} % < 0 < {} %".format(probability1 * 100, probability2 * 100)


### Valence hypothesis testing

# test 1V0A against 0
print(inference_test(v_1V0A.trace()))

# test 1V0A against 0V0A
print(inference_test(v_1V0A.trace() - v_0V0A.trace()))

### Arousal hypothesis testing

# test 0V1A against 0
print(inference_test(v_0V1A.trace()))

# test 0V1A against 0V0A
print(inference_test(v_0V1A.trace() - v_0V0A.trace()))

### Multi-attribute hypothesis testing

# compare 1V1A drift rate magnitude (converted to positive) against 0V1A drift rate magnitude
print(inference_test(-v_1V1A.trace() - v_0V1A.trace()))

# compare 1V1A drift rate magnitude (converted to positive) against 1V0A drift rate magnitude (converted to positive)
print(inference_test(-v_1V1A.trace() - -v_1V0A.trace()))

# compare 1V-1A drift rate magnitude (converted to positive) against 0V1A drift rate magnitude
print(inference_test(-v_1Vne1A.trace() - v_0V1A.trace()))

# compare 1V-1A drift rate magnitude (converted to positive) against 1V0A drift rate magnitude (converted to positive)
print(inference_test(-v_1Vne1A.trace() - -v_1V0A.trace()))


# compare 0V1A drift rate magnitude against 1V0A drift rate magnitude (converted to positive)
print(inference_test(-v_1V0A.trace() - v_0V1A.trace()))

## Credible intervals for drift rate posteriors

# 
stats = m.gen_stats()
stats[stats.index.isin(['v(0V0A)', 'v(1V0A)', 'v(0V1A)', 'v(1V1A)', 'v(1Vne1A)'])]

# test the correlation between drift rate and mood valence/arosual

# load the correlation traces
cor_traces = pd.read_csv("study1_correlation_traces.csv")

# test the correlation against 0 - p values
print(inference_test(cor_traces.cor_valence))
print(inference_test(cor_traces.cor_arousal))

# get the quantiles for the correlation parameter
print(cor_traces.cor_valence.quantile(q=[0.025, 0.5, 0.975]))
print(cor_traces.cor_arousal.quantile(q=[0.025, 0.5, 0.975]))

# Exploratory analysis

# Load the regression model fitted in step XX
m_full_model = hddm.load("m_full_regression_study1_new_subj_10_removed")

# access the sampled node for v, a, t, z
v_intercept, v_movie_valence, v_movie_arousal, v_movie_valence_movie_arousal\
= m_full_model.nodes_db.loc[["v_Intercept",
                             "v_movie_valence[T.1]",
                             "v_movie_arousal[T.1]",
                             "v_movie_valence[T.1]:movie_arousal[T.1]"], 'node']
a_intercept, a_movie_valence, a_movie_arousal, a_movie_valence_movie_arousal\
= m_full_model.nodes_db.loc[["a_Intercept",
                             "a_movie_valence[T.1]",
                             "a_movie_arousal[T.1]",
                             "a_movie_valence[T.1]:movie_arousal[T.1]"], 'node']
t_intercept, t_movie_valence, t_movie_arousal, t_movie_valence_movie_arousal\
= m_full_model.nodes_db.loc[["t_Intercept",
                             "t_movie_valence[T.1]",
                             "t_movie_arousal[T.1]",
                             "t_movie_valence[T.1]:movie_arousal[T.1]"], 'node']
z_intercept, z_movie_valence, z_movie_arousal, z_movie_valence_movie_arousal\
= m_full_model.nodes_db.loc[["z_Intercept",
                             "z_movie_valence[T.1]",
                             "z_movie_arousal[T.1]",
                             "z_movie_valence[T.1]:movie_arousal[T.1]"], 'node']

### testing regressors effects on drift rate

# test the regressors effects against 0
print(inference_test(v_movie_valence.trace()))
print(inference_test(v_movie_arousal.trace()))
print(inference_test(v_movie_valence_movie_arousal.trace()))

### testing regressors effects on decison boundary

print(inference_test(a_movie_valence.trace()))
print(inference_test(a_movie_arousal.trace()))
print(inference_test(a_movie_valence_movie_arousal.trace()))

### testing regressors effects on non-decision time

print(inference_test(t_movie_valence.trace()))
print(inference_test(t_movie_arousal.trace()))
print(inference_test(t_movie_valence_movie_arousal.trace()))

### testing regressors effects on decision bias

print(inference_test(z_movie_valence.trace()))
print(inference_test(z_movie_arousal.trace()))
print(inference_test(z_movie_valence_movie_arousal.trace()))

### credible intervals for parameters: v, a, t, z

stats = m_full_model.gen_stats()
stats[stats.index.isin(["v_Intercept",
    "v_movie_valence[T.1]",
    "v_movie_arousal[T.1]", 
    "v_movie_valence[T.1]:movie_arousal[T.1]",
    
                        "a_Intercept",
    "a_movie_valence[T.1]",
    "a_movie_arousal[T.1]", 
    "a_movie_valence[T.1]:movie_arousal[T.1]",
    
                        "t_Intercept",
    "t_movie_valence[T.1]",
    "t_movie_arousal[T.1]", 
    "t_movie_valence[T.1]:movie_arousal[T.1]",
    
                        "z_Intercept",
    "z_movie_valence[T.1]",
    "z_movie_arousal[T.1]", 
    "z_movie_valence[T.1]:movie_arousal[T.1]",])]

# save the traces for plotting
pd.DataFrame({"v_intercept": v_intercept.trace(),
     "v_movie_valence": v_movie_valence.trace(),
     "v_movie_arousal": v_movie_arousal.trace(), 
     "v_movie_valence_movie_arousal": v_movie_valence_movie_arousal.trace(),
     "a_intercept": a_intercept.trace(),
     "a_movie_valence": a_movie_valence.trace(),
     "a_movie_arousal": a_movie_arousal.trace(),
     "a_movie_valence_movie_arousal": a_movie_valence_movie_arousal.trace(),
     "t_intercept": t_intercept.trace(), 
     "t_movie_valence": t_movie_valence.trace(),
     "t_movie_arousal": t_movie_arousal.trace(),
     "t_movie_valence_movie_arousal": t_movie_valence_movie_arousal.trace(),
     "z_intercept": z_intercept.trace(), 
     "z_movie_valence": z_movie_valence.trace(), 
     "z_movie_arousal": z_movie_arousal.trace(), 
     "z_movie_valence_movie_arousal": z_movie_valence_movie_arousal.trace()}).to_csv("plot/study_1_ddm_exploratory_analysis.csv")
