==> ddm_analysis/step4_Fitting_regression_study_2.ipynb <==
import ipyparallel
import hddm
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

print(hddm.__version__)

### Load the data

# read the data to pandas dataframe
data = hddm.load_csv('data/study2.csv')

# flip RTs of choosing lower value options to negative RTs
data = hddm.utils.flip_errors(data)
data

# code the induction group of each subject into 2*2 between-subject variables
# mood_valence & mood_arousal
# code the decision tasks into 2*2 within-subject variables
# movie_valence & movie_arousal

data = data.drop(data[data.stim == '1Vne1A'].index)
data['movie_valence'] = data.stim.str[:1]
data['movie_arousal'] = data.stim.str[-2:-1]
di = {'group1': "1V1A", 'group2': "1V0A", 'group3': "0V1A", 'group4': "0V0A"}
data['group_code'] = data.group
data = data.replace({"group": di})
data['mood_valence'] = data.group.str[:1]
data['mood_arousal'] = data.group.str[-2:-1]

data

# FULL Regression Model Fixed Model

# sepecify the HDDM regression model
# include four decision parameters: 
# drift rate: v; decision boundary: a; non-decision time: t; decision bias: z
# specify mixture model with 5% of the RTs as contations RT from a uniform distribution

m_FULL_hypothesis = hddm.HDDMRegressor(data, ["v ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "a ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "t ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "z ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal"],
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

# find initial fitting estimate for each parameter
# speed up MCMC chain fitting
m_FULL_hypothesis.find_starting_values()

# start sampling MCMC chains
# start drawing 10000 samples, with 2000 as burn-in samples
# save the fitting data into a bd file
m_FULL_hypothesis.sample(10000, burn=2000, dbname='study2_regression_full_hypothesis.db', db='pickle')

# save the model
m_FULL_hypothesis.save('study2_regression_full_hypothesis')

### Fixed effect interaction model

# interaction model gives effect of movie_valence at different level of mood valence
# and effect of movie arousal at different level of mood arousal
# note: this model is equivalent to m_FULL_hypothesis
# check pasty documentation for more details
# https://patsy.readthedocs.io/en/latest/

m_FULL_hypothesis_interaction = hddm.HDDMRegressor(data, ["v ~ movie_valence:mood_valence + movie_arousal:mood_arousal",
                                   "a ~ movie_valence:mood_valence + movie_arousal:mood_arousal",
                                   "t ~ movie_valence:mood_valence + movie_arousal:mood_arousal",
                                   "z ~ movie_valence:mood_valence + movie_arousal:mood_arousal"],
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

# fit the model and save the model
m_FULL_hypothesis_interaction.find_starting_values()
m_FULL_hypothesis_interaction.sample(10000, burn=2000, dbname='study2_regression_full_hypothesis_interaction.db', db='pickle')
m_FULL_hypothesis_interaction.save('study2_regression_full_hypothesis_interaction')

# Reduced models
### We fitted several reduced models, and tested if the reduced model has better fit with our empirical data.

# reduced model with only drift rate
m_FULL_hypothesis_only_v = hddm.HDDMRegressor(data, ["v ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"],
                            include=('v'),
                            p_outlier=0.05)

# reduced model with only drift rate and decision boundary
m_FULL_hypothesis_only_v_a = hddm.HDDMRegressor(data, ["v ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "a ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal"],
                            include=('v', 'a'),
                            p_outlier=0.05)

# reduced model with only drift rate, decision boundary, and non-decision time
m_FULL_hypothesis_only_v_a_t = hddm.HDDMRegressor(data, ["v ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "a ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal",
                                   "t ~ movie_valence + movie_arousal + mood_valence + mood_arousal"
                                   + "+ movie_valence * mood_valence"
                                   + "+ movie_arousal * mood_arousal"],
                            include=('v', 'a', 't'),
                            p_outlier=0.05)

# reduced model with only valence
m_FULL_hypothesis_valence_only = hddm.HDDMRegressor(data, ["v ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "a ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "t ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "z ~ movie_valence + mood_valence + movie_valence * mood_valence"],
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

# reduced model with only arousal
m_FULL_hypothesis_valence_only = hddm.HDDMRegressor(data, ["v ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "a ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "t ~ movie_valence + mood_valence + movie_valence * mood_valence",
                                   "z ~ movie_valence + mood_valence + movie_valence * mood_valence"],
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

m_FULL_hypothesis_arousal_only = hddm.HDDMRegressor(data, ["v ~ movie_arousal + mood_arousal + movie_arousal * mood_arousal",
                                   "a ~  movie_arousal + mood_arousal + movie_arousal * mood_arousal",
                                   "t ~  movie_arousal + mood_arousal + movie_arousal * mood_arousal",
                                   "z ~  movie_arousal + mood_arousal + movie_arousal * mood_arousal"],
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

# fit all the reduced model and save the model
m_FULL_hypothesis_only_v.find_starting_values()
m_FULL_hypothesis_only_v.sample(10000, burn=2000, dbname='traces.db', db='pickle')
m_FULL_hypothesis_only_v.save('m_FULL_hypothesis_only_v')

m_FULL_hypothesis_only_v_a.find_starting_values()
m_FULL_hypothesis_only_v_a.sample(10000, burn=2000, dbname='traces.db', db='pickle')
m_FULL_hypothesis_only_v_a.save('m_FULL_hypothesis_only_v_a')

m_FULL_hypothesis_only_v_a_t.find_starting_values()
m_FULL_hypothesis_only_v_a_t.sample(10000, burn=2000, dbname='traces.db', db='pickle')
m_FULL_hypothesis_only_v_a_t.save('m_FULL_hypothesis_only_v_a_t')

m_FULL_hypothesis_valence_only.find_starting_values()
m_FULL_hypothesis_valence_only.sample(10000, burn=2000, dbname='traces.db', db='pickle')
m_FULL_hypothesis_valence_only.save('m_FULL_hypothesis_valence_only')

m_FULL_hypothesis_arousal_only.find_starting_values()
m_FULL_hypothesis_arousal_only.sample(10000, burn=2000, dbname='traces.db', db='pickle')
m_FULL_hypothesis_arousal_only.save('m_FULL_hypothesis_arousal_only')

## Compare the DIC value of all the models
### Lower dic value indicate better model fitting

print(m_FULL_hypothesis.dic)
print(m_FULL_hypothesis_only_v.dic)
print(m_FULL_hypothesis_only_v_a.dic)
print(m_FULL_hypothesis_only_v_a_t.dic)
print(m_FULL_hypothesis_valence_only.dic)
print(m_FULL_hypothesis_arousal_only.dic)
