==> ddm_analysis/step6_Cell_means_study_2.ipynb <==
import hddm
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
%matplotlib inline

print(hddm.__version__)

# load the data

data = hddm.load_csv('data/study2.csv')

data = hddm.utils.flip_errors(data)
data

data = data.drop(data[data.stim == '1Vne1A'].index)
data['movie_valence'] = data.stim.str[:1]
data['movie_arousal'] = data.stim.str[-2:-1]
di = {'group1': "1V1A", 'group2': "1V0A", 'group3': "0V1A", 'group4': "0V0A"}
data['group_code'] = data.group
data = data.replace({"group": di})
data['mood_valence'] = data.group.str[:1]
data['mood_arousal'] = data.group.str[-2:-1]

# fit a HDDM with parameters depends on the 4 variables

### 2 (movie_valence) * 2 (movie_arousal) * 2 (mood_valence) * 2 (mood_arousal)

m_cell_means_all = hddm.HDDM(data, depends_on={'v': ['stim', 'group'],
                                          'a': ['stim', 'group'],
                                          't': ['stim', 'group'],
                                          'z': ['stim', 'group']},
                            include=('v', 'a', 't', 'z'),
                            p_outlier=0.05)

m_cell_means_all.find_starting_values()
m_cell_means_all.sample(10000, burn=2000, dbname='study2_cell_mean_all.db', db='pickle')
m_cell_means_all.save('study2_cell_mean_all')

v_0V0A_0V0A, v_0V0A_0V1A, v_0V0A_1V0A, v_0V0A_1V1A, v_1V0A_0V0A, v_1V0A_0V1A, v_1V0A_1V0A, v_1V0A_1V1A, \
v_0V1A_0V0A, v_0V1A_0V1A, v_0V1A_1V0A, v_0V1A_1V1A, v_1V1A_0V0A, v_1V1A_0V1A, v_1V1A_1V0A, v_1V1A_1V1A, \
a_0V0A_0V0A, a_0V0A_0V1A, a_0V0A_1V0A, a_0V0A_1V1A, a_1V0A_0V0A, a_1V0A_0V1A, a_1V0A_1V0A, a_1V0A_1V1A, \
a_0V1A_0V0A, a_0V1A_0V1A, a_0V1A_1V0A, a_0V1A_1V1A, a_1V1A_0V0A, a_1V1A_0V1A, a_1V1A_1V0A, a_1V1A_1V1A, \
t_0V0A_0V0A, t_0V0A_0V1A, t_0V0A_1V0A, t_0V0A_1V1A, t_1V0A_0V0A, t_1V0A_0V1A, t_1V0A_1V0A, t_1V0A_1V1A, \
t_0V1A_0V0A, t_0V1A_0V1A, t_0V1A_1V0A, t_0V1A_1V1A, t_1V1A_0V0A, t_1V1A_0V1A, t_1V1A_1V0A, t_1V1A_1V1A, \
z_0V0A_0V0A, z_0V0A_0V1A, z_0V0A_1V0A, z_0V0A_1V1A, z_1V0A_0V0A, z_1V0A_0V1A, z_1V0A_1V0A, z_1V0A_1V1A, \
z_0V1A_0V0A, z_0V1A_0V1A, z_0V1A_1V0A, z_0V1A_1V1A, z_1V1A_0V0A, z_1V1A_0V1A, z_1V1A_1V0A, z_1V1A_1V1A = m_cell_means_all.nodes_db.loc[[
    "v(0V0A.0V0A)",
    "v(0V0A.0V1A)",
    "v(0V0A.1V0A)",
    "v(0V0A.1V1A)",
    "v(1V0A.0V0A)",
    "v(1V0A.0V1A)",
    "v(1V0A.1V0A)",
    "v(1V0A.1V1A)",
    "v(0V1A.0V0A)",
    "v(0V1A.0V1A)",
    "v(0V1A.1V0A)",
    "v(0V1A.1V1A)",
    "v(1V1A.0V0A)",
    "v(1V1A.0V1A)",
    "v(1V1A.1V0A)",
    "v(1V1A.1V1A)",
    
    "a(0V0A.0V0A)",
    "a(0V0A.0V1A)",
    "a(0V0A.1V0A)",
    "a(0V0A.1V1A)",
    "a(1V0A.0V0A)",
    "a(1V0A.0V1A)",
    "a(1V0A.1V0A)",
    "a(1V0A.1V1A)",
    "a(0V1A.0V0A)",
    "a(0V1A.0V1A)",
    "a(0V1A.1V0A)",
    "a(0V1A.1V1A)",
    "a(1V1A.0V0A)",
    "a(1V1A.0V1A)",
    "a(1V1A.1V0A)",
    "a(1V1A.1V1A)",
    
    "t(0V0A.0V0A)",
    "t(0V0A.0V1A)",
    "t(0V0A.1V0A)",
    "t(0V0A.1V1A)",
    "t(1V0A.0V0A)",
    "t(1V0A.0V1A)",
    "t(1V0A.1V0A)",
    "t(1V0A.1V1A)",
    "t(0V1A.0V0A)",
    "t(0V1A.0V1A)",
    "t(0V1A.1V0A)",
    "t(0V1A.1V1A)",
    "t(1V1A.0V0A)",
    "t(1V1A.0V1A)",
    "t(1V1A.1V0A)",
    "t(1V1A.1V1A)",
    
    "z(0V0A.0V0A)",
    "z(0V0A.0V1A)",
    "z(0V0A.1V0A)",
    "z(0V0A.1V1A)",
    "z(1V0A.0V0A)",
    "z(1V0A.0V1A)",
    "z(1V0A.1V0A)",
    "z(1V0A.1V1A)",
    "z(0V1A.0V0A)",
    "z(0V1A.0V1A)",
    "z(0V1A.1V0A)",
    "z(0V1A.1V1A)",
    "z(1V1A.0V0A)",
    "z(1V1A.0V1A)",
    "z(1V1A.1V0A)",
    "z(1V1A.1V1A)",], 'node']

## Create a dataframe includes the drift rate samples as DV and the IVs

trace = np.concatenate((v_0V0A_0V0A.trace(), v_0V0A_0V1A.trace(), v_0V0A_1V0A.trace(), v_0V0A_1V1A.trace(),
               v_1V0A_0V0A.trace(), v_1V0A_0V1A.trace(), v_1V0A_1V0A.trace(), v_1V0A_1V1A.trace(), \
               v_0V1A_0V0A.trace(), v_0V1A_0V1A.trace(), v_0V1A_1V0A.trace(), v_0V1A_1V1A.trace(), \
               v_1V1A_0V0A.trace(), v_1V1A_0V1A.trace(), v_1V1A_1V0A.trace(), v_1V1A_1V1A.trace()))

mood = np.concatenate((np.repeat("0V0A", 8000), np.repeat("0V0A", 8000), np.repeat("0V0A", 8000), np.repeat("0V0A", 8000),
               np.repeat("1V0A", 8000), np.repeat("1V0A", 8000), np.repeat("1V0A", 8000), np.repeat("1V0A", 8000),
               np.repeat("0V1A", 8000), np.repeat("0V1A", 8000), np.repeat("0V1A", 8000), np.repeat("0V1A", 8000),
               np.repeat("1V1A", 8000), np.repeat("1V1A", 8000), np.repeat("1V1A", 8000), np.repeat("1V1A", 8000)))

mood_valence = np.concatenate((np.repeat("negative valence", 8000*4), 
               np.repeat("positive valence", 8000*4),
               np.repeat("negative valence", 8000*4),
               np.repeat("positive valence", 8000*4)))

mood_arousal = np.concatenate((np.repeat("low arousal", 8000*8), 
               np.repeat("high arousal", 8000*8)))

movie = np.concatenate((np.repeat("0V0A", 8000), np.repeat("0V1A", 8000), np.repeat("1V0A", 8000), np.repeat("1V1A", 8000),
               np.repeat("0V0A", 8000), np.repeat("0V1A", 8000), np.repeat("1V0A", 8000), np.repeat("1V1A", 8000),
               np.repeat("0V0A", 8000), np.repeat("0V1A", 8000), np.repeat("1V0A", 8000), np.repeat("1V1A", 8000),
               np.repeat("0V0A", 8000), np.repeat("0V1A", 8000), np.repeat("1V0A", 8000), np.repeat("1V1A", 8000)))

movie_valence = np.concatenate((np.repeat("negative valence", 8000*2), np.repeat("positive valence", 8000*2),
               np.repeat("negative valence", 8000*2), np.repeat("positive valence", 8000*2),
               np.repeat("negative valence", 8000*2), np.repeat("positive valence", 8000*2),
               np.repeat("negative valence", 8000*2), np.repeat("positive valence", 8000*2)))

movie_arousal = np.tile(np.concatenate((np.repeat("low arousal", 8000), np.repeat("high arousal", 8000))),8)

trace_df = pd.DataFrame({"trace": trace,
                        "movie": movie,
                        "mood": mood,
                        "movie_valence": movie_valence,
                        "movie_arousal": movie_arousal,
                        "mood_valence": mood_valence,
                        "mood_arousal": mood_arousal})

# save the dataframe for plotting
trace_df.to_csv("plot/study2_cell_mean_trace_df.csv")
